+++ dbms/src/Flash/Mpp/FineGrainedShuffleWriter.cpp
@@ -146,7 +146,7 @@ void FineGrainedShuffleWriter<StreamWriterPtr>::writePackets(std::vector<Tracked
     if constexpr (send_exec_summary_at_last)
     {
         tipb::SelectResponse response;
-        addExecuteSummaries(response, /*delta_mode=*/false);
+        summary_collector.addExecuteSummaries(response, /*delta_mode=*/false);
         /// Sending the response to only one node, default the first one.
         assert(!packets.empty());
         packets[0].serializeByResponse(response);
